requirements.txt
black==22.3.0
click==8.1.2
cycler==0.11.0
fonttools==4.31.2
kiwisolver==1.4.2
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.8.0
six==1.16.0
tomli==2.0.1
